Complete refactoring, from a prime example of overengineering
the LOC is down from 8748 -> 1004, while maintaining same level of
code coverage & retaining 90% of functionality (only caching
was removed).

Caching was dropped as the expected use cases are mostly in CI,
checking for 1-10? releases so caching would hardly ever
be useful.

Also, added a bit of colors. And a command to generate a SVG
badge for a product release (appropriately color coded based
on EOL, EOAS or MAINTAINED).